+++ data-transfer-object/src/main/java/org/grnet/status/dtos/tenant/TenantRequestDto.java
@@ -60,4 +60,14 @@ public class TenantRequestDto {
     )
     @JsonProperty("performance")
     public Boolean performance;
+
+    @Schema(
+            type = SchemaType.BOOLEAN,
+            implementation = Boolean.class,
+            description = "Indicates whether the tenant is configured to have public downtimes",
+            example = "true"
+    )
+    @JsonProperty("public_downtime")
+    public Boolean publicDowntime;
+
 }