+++ entity/src/main/resources/db/migration/tables/tenant/V1.22__update_tenant_downtime.sql
@@ -1,0 +1,7 @@
+-- ------------------------------------------------
+-- Version: V1.22
+-- Description: Add publicDowntime flag to t_Tenant
+-- ------------------------------------------------
+
+ALTER TABLE t_Tenant
+    ADD COLUMN public_downtime BOOLEAN NOT NULL DEFAULT FALSE;